anthropic-agent/requirements.txt
khaos-agent>=1.0.0
anthropic>=0.30
